chore: Update packages and add TypeScript files
This commit updates the package-lock.json and package.json files to include new dependencies for TypeScript. It adds "@types/jest", "@types/node", "@types/react", and "@types/react-dom" with their respective versions. Additionally, it updates the "typescript" dependency to version 5.2.2.

In the source code, it removes the App.js file and adds a new App.tsx file. The new file imports the App.css file and exports a functional component that renders a div with the text "Hello world!".

Lastly, it renames the index.js file to index.tsx and adds a new tsconfig.json file.